requirements.txt
Pillow==11.0.0
requests==2.27.0
rapidfuzz==3.1.1
python-dotenv==1.0.0